정리

+++ database/migrations/2020_07_11_061355_create_foreign_table.php
@@ -40,11 +40,11 @@ class CreateForeignTable extends Migration
               $table->integer('chat_num')->index();
               $table->foreign('chat_num')->references('chatroom_num')->on('chatroom');
             });
-            Schema::table('comment',function (Blueprint $table){
+            Schema::table('comment', function (Blueprint $table){
               $table->string('comment_id')->index();
               $table->foreign('comment_id')->references('id')->on('users');
-              
-              $table->integer('comm_item')->index();
+
+              $table->unsignedBigInteger('comm_item')->index();
               $table->foreign('comm_item')->references('item_number')->on('items');
             });
     }
